gbms: misc fixes for testing
No functional changes, only fix compile issues for patchwork.

Bug: 213492076

+++ gbms_power_supply.h
@@ -101,7 +101,7 @@ static inline int gpsy_set_int64_prop(struct power_supply *psy,
 		return -EINVAL;
 
 	pr_debug("set %s for '%s' to %lld\n", prop_name,
-		 psy->desc->name, val.int64val);
+		 psy->desc->name, (long long)val.int64val);
 
 	ret = power_supply_set_property(psy, (enum power_supply_property)psp,
 				        &val.prop);
@@ -137,7 +137,7 @@ static inline int64_t gpsy_get_int64_prop(struct power_supply *psy,
 	}
 
 	pr_debug("get %s for '%s' => %lld\n", prop_name,
-		 psy->desc->name, val.int64val);
+		 psy->desc->name, (long long)val.int64val);
 
 	return val.int64val;
 }
